% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2GTFS.R
\name{sf2GTFS}
\alias{sf2GTFS}
\title{Download GTFS files from a shapefile}
\usage{
sf2GTFS(urls, files, output_dir)
}
\arguments{
\item{urls}{A character vector of URLs to download GTFS files from.}

\item{files}{A character vector of filenames corresponding to each URL.}

\item{output_dir}{A character string specifying the directory to save the downloaded files.}

\item{api_key}{Optional API key for accessing restricted URLs.}
}
\description{
This function downloads GTFS files from a specified shapefile and saves them to a specified directory.
}
