% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2GTFS.R
\name{sf2GTFS}
\alias{sf2GTFS}
\title{Download static GTFS feeds from a simple features object}
\usage{
sf2GTFS(sf, output_dir)
}
\arguments{
\item{output_dir}{A character string specifying the directory to save the downloaded files.}

\item{sf_object}{An sf class object representing the area(s) of interest.}
}
\description{
This function downloads static GTFS feeds for all transit routes operating
within all of the geographies of a simple features (sf) class object and saves them to directory.
}
